+++ lib/jobmanager.py
@@ -7,9 +7,8 @@ import qpack
 from enodo.jobs import *
 from enodo.model.config.series import SeriesJobConfigModel
 from enodo.net import PROTO_REQ_WORKER_REQUEST
-from enodo.protocol.packagedata import (
-    EnodoJobDataModel, EnodoRequest, EnodoRequestResponse,
-    REQUEST_TYPE_EXTERNAL)
+from enodo.protocol.packagedata import (EnodoRequest, EnodoRequestResponse,
+                                        REQUEST_TYPE_EXTERNAL)
 from lib.outputmanager import EnodoOutputManager
 from lib.socket.clientmanager import WorkerClient
 from .socket import ClientManager
@@ -25,20 +24,15 @@ class EnodoJob:
                  series_name: str,
                  job_config: SeriesJobConfigModel,
                  pool_idx: int,
-                 job_data: Optional[dict] = None,
                  send_at: Optional[int] = None,
                  error: Optional[str] = None,
                  worker_id: Optional[str] = None):
-        if not isinstance(
-                job_data, EnodoJobDataModel) and job_data is not None:
-            raise Exception('Unknown job data value')
         if rid is None:
             rid = str(uuid4()).replace("-", "")
         self.rid = rid
         self.series_name = series_name
         self.job_config = job_config
         self.pool_idx = pool_idx
-        self.job_data = job_data
         self.send_at = send_at
         self.error = error
         self.worker_id = worker_id